+++ CHANGELOG.md
@@ -29,6 +29,7 @@ This change means that a lot of boilerplate code can be removed. Also keep in mi
 - `State::isOneOf()` is removed, `State::equals` now accepts multiple state objects.
 - `State::is()` is removed, you should use `State::equals()`.
 - Dropped support for Laravel 5, 6, and 7. The minimal required version is `laravel/framework:^8`
+- Proper support for `finalState` in `StateChanged` event
 
 ## 1.9.0 - 2020-08-24